requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyairtable>=2.3.3
markdownify>=0.13.1
